removing diagnostics files and adding to generic driver
