+++ emulator/drivers/burnwire.py
@@ -1,6 +1,12 @@
-class BurnWires:
-    def __init__(self) -> None:
-        pass
+from hal.drivers.middleware.generic_driver import Driver
+
+
+class BurnWires(Driver):
+    def __init__(self, enable=None) -> None:
+        super().__init__(enable)
 
     def duty_cycle(self, duty_cycle):
         assert 0 <= duty_cycle <= 0xFFFF
+
+    def run_diagnostics(self) -> list:
+        return {}
